Music app: Fix for the issue of playback cursor stopping one sec before final duration.
Music playback activity refreshes UI for every one second. If the music clip is
19.8 seconds, the last refresh for the clip is at 19th second, and next refresh
will show the status of the next music clip. Because of this UI shows as if the
clip is being played for 19 seconds only.

Fixed the issue by refreshing UI when the playback of a clip is completed.

CRs-fixed: 189780