Music : Fix for crash in artistalbum browser observed when sdcard gets unmounted
While in ArtistlbumBrowser, if sdcard gets unmounted, list of tracks is
still being displayed and leads to a crash when any of the items in the
list is clicked. This is happening because on receiving the unmounted
event, the changeCursor() needs to update the mArtsistCursor even when the cursor
is null.  Hence, modified the condition to update the same.
Similar change needs to be made in AlbumBrowser Activity.

CRs-Fixed: 212568