platform/packages/apps/Music: Rewind on first clip at last second jumps to middle of next clip
The seek request which is generated in the first clip at the last second
of play is processed in the next clip. This leads to second clip not
being played from the beginning. The fix is to skip the seek request
generated in the last second of the first clip.

CRs-Fixed: 212306